Allow mail to be embedded into a Jinja2 template
If template_dir is specified in the config file, muttdown will use
the Jinja2 file named by config.template (default: muttdown.tmpl) in
template_dir to render processed parts.

There are two content variables supplied to the template:

- body
- signature

This can be used to provide a body/table container for the generated
email.

The .md source can remain clean and the HTML email can have styles
attached to the container.